moving download logic into service
