+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.component.html
@@ -86,7 +86,7 @@
     </div>
 
     <div class="save-btn">
-        <md-button class="md-fab md-primary dark_shadow_opacity" ng-click="$ctrl.prepareFileDownload()">
+        <md-button class="md-fab md-primary dark_shadow_opacity" ng-click="$ctrl.downloadFile()">
             <i class="fa fa-download"></i>
         </md-button>
     </div>
