+++ visualization/app/codeCharta/ui/ribbonBar/ribbonBar.component.ts
@@ -1,6 +1,7 @@
 import "./ribbonBar.component.scss";
 import $ from "jquery";
 import {SettingsService} from "../../core/settings/settings.service";
+import { DownloadService } from "../../core/download/download.service";
 
 export class RibbonBarController {
 
@@ -10,8 +11,13 @@ export class RibbonBarController {
 
     /* @ngInject */
     constructor(
-        private settingsService: SettingsService
+        private settingsService: SettingsService,
+        private downloadService: DownloadService
     ) {
+    }
+
+    public downloadFile() {
+        this.downloadService.prepareFileDownload();
     }
 
     public changeMargin(){
@@ -37,55 +43,6 @@ export class RibbonBarController {
         this.collapsingElements.removeClass("expanded");
     }
 
-    addDateToFileName(fileName) {
-        const date = new Date();
-        const dateString = date.getDate() + "_" + (date.getMonth() + 1)  + "_" + date.getFullYear();
-        let tokens = fileName.split(".");
-        tokens.splice(1, 0, dateString);
-        return tokens.join(".");
-    }
-
-    addJsonFileEndingIfNecessary(fileName) {
-        if(!fileName.endsWith(".json")) {
-            return fileName + ".json";
-        }
-        return fileName;
-    }
-
-    prepareFileDownload() {
-        var settings: any = this.settingsService.settings;
-        var map: any = settings.map;
-
-        const datedFileName = this.addDateToFileName(map.fileName);
-        const resultingFileName = this.addJsonFileEndingIfNecessary(datedFileName);
-
-        let data = {
-            fileName: resultingFileName,
-            projectName: map.projectName,
-            apiVersion: map.apiVersion,
-            nodes: [map.root],
-            edges: map.edges,
-            attributeTypes: map.attributeTypes,
-            blacklist: settings.blacklist,
-        };
-        this.downloadData(data, resultingFileName);
-    }
-
-    downloadData(data, fileName) {
-        if (typeof data === "object") {
-            data = JSON.stringify(data, undefined, 4);
-        }
-
-        const blob = new Blob([data], {type: "text/json"});
-        const e = document.createEvent("MouseEvents");
-        const a = document.createElement("a");
-
-        a.download = fileName;
-        a.href = window.URL.createObjectURL(blob);
-        a.dataset.downloadurl = ["text/json", a.download, a.href].join(":");
-        e.initMouseEvent("click", true, false, window, 0, 0, 0, 0, 0, false, false, false, false, 0, null);
-        a.dispatchEvent(e);
-    }
 }
 
 export const ribbonBarComponent = {
